refactor: extract FeedImageDataStory Spy into shared helper scope
