+++ EssentialFeed/EssentialFeedTests/Feed Cache/Helpers/FeedImageDataStorySpy.swift
@@ -1,0 +1,37 @@
+//
+//  FeedImageDataStorySpy.swift
+//  EssentialFeedTests
+//
+//  Created by Nikhil Menon on 4/17/24.
+//
+
+import Foundation
+import EssentialFeed
+
+class FeedImageDataStoreSpy: FeedImageDataStore {
+    enum Message: Equatable {
+        case retrieve(dataFor: URL)
+        case insert(data: Data, for: URL)
+    }
+
+    private var retrievalCompletions = [(FeedImageDataStore.RetrievalResult) -> Void]()
+    
+    private(set) var receivedMessages = [Message]()
+    
+    func retrieve(dataForURL url: URL, completion: @escaping (FeedImageDataStore.RetrievalResult) -> Void) {
+        receivedMessages.append(.retrieve(dataFor: url))
+        retrievalCompletions.append(completion)
+    }
+    
+    func insert(_ data: Data, for url: URL, completion: @escaping (FeedImageDataStore.InsertionResult) -> Void) {
+        receivedMessages.append(.insert(data: data, for: url))
+    }
+    
+    func complete(with error: Error, at index: Int = 0) {
+        retrievalCompletions[index](.failure(error))
+    }
+    
+    func complete(with data: Data?, at index: Int = 0) {
+        retrievalCompletions[index](.success(data))
+    }
+}
